requirements.txt
alabaster==0.7.12
asgiref==3.4.1
attrs==21.4.0
Babel==2.9.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.10
choicesenum==0.7.0
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.8
django-rest-auth==0.9.5
djangorestframework==3.13.1
docutils==0.17.1
drf-spectacular==0.21.1
drf-spectacular-sidecar==2021.12.13
idna==3.3
imagesize==1.3.0
importlib-metadata==4.10.1
importlib-resources==5.4.0
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.0.3
jsondiff==1.3.1
jsonschema==4.4.0
MarkupSafe==2.0.1
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
Pygments==2.11.2
pyparsing==3.0.7
pyrsistent==0.18.1
pytest==6.2.5
pytest-django==4.5.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
six==1.16.0
snowballstemmer==2.2.0
Sphinx==4.4.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.8
zipp==3.7.0
